Implemented TextArea

+++ src/lwjgui/LWJGUIUtil.java
@@ -67,14 +67,14 @@ public class LWJGUIUtil {
 
 	public static void fillRect(Context context, double x, double y, double width, double height, Color color) {
 		NanoVG.nvgBeginPath(context.getNVG());
-		NanoVG.nvgRect(context.getNVG(), (int)x, (int)y, (float)width, (float)height);
+		NanoVG.nvgRect(context.getNVG(), (int)x, (int)y, (int)width, (int)height);
 		NanoVG.nvgFillColor(context.getNVG(), color.getNVG());
 		NanoVG.nvgFill(context.getNVG());
 	}
 
 	public static void fillRoundRect(Context context, double x, double y, double width, double height, double radius, Color color) {
 		NanoVG.nvgBeginPath(context.getNVG());
-		NanoVG.nvgRoundedRect(context.getNVG(), (float)x, (float)y, (float)width, (float)height, (float)radius);
+		NanoVG.nvgRoundedRect(context.getNVG(), (int)x, (int)y, (int)width, (int)height, (float)radius);
 		NanoVG.nvgFillColor(context.getNVG(), color.getNVG());
 		NanoVG.nvgFill(context.getNVG());
 	}
